Add changed based on PR review: moving things around and improving docs
