+++ distributed/diagnostics/__init__.py
@@ -1,6 +1,6 @@
 from ..utils import ignoring
 from .graph_layout import GraphLayout
-from .scheduler_plugin import SchedulerPlugin
+from .plugin import SchedulerPlugin
 
 with ignoring(ImportError):
     from .progressbar import progress
